requirements.txt
alembic==1.7.5
anyio==3.5.0
asgiref==3.4.1
bcrypt==3.2.0
certifi==2021.10.8
cffi==1.15.0
click==8.0.3
cloudinary==1.28.1
colorama==0.4.4
cryptography==36.0.1
ecdsa==0.17.0
fastapi==0.71.0
greenlet==1.1.2
gunicorn==20.1.0
h11==0.12.0
httptools==0.3.0
idna==3.3
Mako==1.1.6
MarkupSafe==2.0.1
passlib==1.7.4
psycopg2==2.9.3
pyasn1==0.4.8
pycparser==2.21
pydantic==1.9.0
python-jose==3.3.0
python-multipart==0.0.5
rsa==4.8
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.29
SQLAlchemy-Utils==0.38.2
starlette==0.17.1
typing_extensions==4.0.1
urllib3==1.26.8
uvicorn==0.16.0
